# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PayD Backend API",
    "version": "1.0.0",
    "description": "API Documentation for PayD Backend"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/payroll-bonus/runs": {
      "post": {
        "summary": "Create a new payroll bonus run",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "summary": "List all payroll bonus runs",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/runs/{id}": {
      "get": {
        "summary": "Get details of a payroll bonus run",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/runs/{id}/status": {
      "patch": {
        "summary": "Update status of a payroll bonus run",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/runs/{id}/execute": {
      "post": {
        "summary": "Execute a payroll bonus run (asynchronous)",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/items/bonus": {
      "post": {
        "summary": "Add a single bonus item",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/items/bonus/batch": {
      "post": {
        "summary": "Add multiple bonus items in batch",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/runs/{payrollRunId}/items": {
      "get": {
        "summary": "List all items in a payroll bonus run",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "payrollRunId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/items/{itemId}": {
      "delete": {
        "summary": "Delete a payroll item",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/bonuses/history": {
      "get": {
        "summary": "Get history of all bonuses",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/bonuses/by-type/{bonusType}": {
      "get": {
        "summary": "List bonus items filtered by type (performance, referral, project, etc.)",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bonusType",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "performance",
                "referral",
                "project",
                "retention",
                "spot",
                "other"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid bonus type"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/bonuses/performance": {
      "get": {
        "summary": "List performance bonuses, optionally filtered by minimum score",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "minScore",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100
            },
            "description": "Minimum performance score (0–100)"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid minScore"
          }
        }
      }
    },
    "/api/v1/payroll-bonus/runs/{id}/report": {
      "get": {
        "summary": "Get a combined payroll run report (run metadata + full audit trail)",
        "tags": [
          "Payroll Bonus"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Payroll run ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Audit log page (default 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Audit log page size (default 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "Combined run summary and audit trail"
          },
          "404": {
            "description": "Payroll run not found"
          }
        }
      }
    },
...
  }
}